# perf(logqlengine): use simple prefix tree to match JSON paths