--- a/internal/logql/logqlengine/jsonexpr/eval.go
+++ b/internal/logql/logqlengine/jsonexpr/eval.go
@@ -11,125 +11,128 @@ import (
 // Extract extracts values from given paths using extract callback.
 func Extract(
 	d *jx.Decoder,
-	paths map[logql.Label]Path,
+	paths SelectorTree,
 	extract func(logql.Label, string),
 ) error {
 	e := extractor{
-		// TODO(tdakkota): re-use slice?
-		current: make(Path, 0, 16),
 		paths:   paths,
 		extract: extract,
 	}
-	return e.walk(d)
+	return e.walk(e.paths.root, d)
 }
 
 // extractor extracts values from JSON value.
 type extractor struct {
-	current Path
-	paths   map[logql.Label]Path
+	paths   SelectorTree
 	extract func(logql.Label, string)
 }
 
-func (e *extractor) walk(d *jx.Decoder) error {
+func (e *extractor) walk(n *selectorNode, d *jx.Decoder) error {
 	switch d.Next() {
 	case jx.String:
 		val, err := d.Str()
 		if err != nil {
 			return err
 		}
-		e.matchLiteral(val)
+		e.matchLiteral(n, val)
 		return nil
 	case jx.Number:
 		num, err := d.Num()
 		if err != nil {
 			return err
 		}
-		e.matchLiteral(string(num))
+		e.matchLiteral(n, string(num))
 		return nil
 	case jx.Null:
 		if err := d.Null(); err != nil {
 			return err
 		}
-		e.matchLiteral("")
+		e.matchLiteral(n, "")
 		return nil
 	case jx.Bool:
 		val, err := d.Bool()
 		if err != nil {
 			return err
 		}
-		e.matchLiteral(strconv.FormatBool(val))
+		e.matchLiteral(n, strconv.FormatBool(val))
 		return nil
 	case jx.Array:
-		return e.walkArr(d)
+		return e.walkArr(n, d)
 	case jx.Object:
-		return e.walkObj(d)
+		return e.walkObj(n, d)
 	default:
 		return d.Skip()
 	}
 }
 
-func (e *extractor) walkObj(d *jx.Decoder) error {
-	if err := e.tryMatchRaw(d); err != nil {
+func (e *extractor) walkObj(n *selectorNode, d *jx.Decoder) error {
+	if err := e.tryMatchRaw(n, d); err != nil {
 		return err
 	}
 
 	return d.Obj(func(d *jx.Decoder, key string) error {
-		e.current = append(e.current, KeySel(key))
-		if err := e.walk(d); err != nil {
+		child, ok := n.sub[KeySel(key)]
+		if !ok {
+			return d.Skip()
+		}
+
+		if err := e.walk(child, d); err != nil {
 			return err
 		}
-		e.current = e.current[:len(e.current)-1]
 		return nil
 	})
 }
 
-func (e *extractor) walkArr(d *jx.Decoder) error {
-	if err := e.tryMatchRaw(d); err != nil {
+func (e *extractor) walkArr(n *selectorNode, d *jx.Decoder) error {
+	if err := e.tryMatchRaw(n, d); err != nil {
 		return err
 	}
 
-	n := 0
+	c := 0
 	return d.Arr(func(d *jx.Decoder) error {
-		e.current = append(e.current, IndexSel(n))
-		if err := e.walk(d); err != nil {
+		child, ok := n.sub[IndexSel(c)]
+		if !ok {
+			return d.Skip()
+		}
+
+		if err := e.walk(child, d); err != nil {
 			return err
 		}
-		e.current = e.current[:len(e.current)-1]
-		n++
+		c++
 		return nil
 	})
 }
 
-func (e *extractor) matchLiteral(val string) {
-	for label, p := range e.paths {
-		if e.current.Equal(p) {
-			e.extract(label, val)
-		}
+func (e *extractor) matchLiteral(n *selectorNode, val string) {
+	for _, label := range n.labels {
+		e.extract(label, val)
 	}
 }
 
-func (e *extractor) tryMatchRaw(d *jx.Decoder) error {
-	var raw string
-	for label, p := range e.paths {
-		if !e.current.Equal(p) {
-			// No match.
-			continue
-		}
+func (e *extractor) tryMatchRaw(n *selectorNode, d *jx.Decoder) error {
+	if len(n.labels) == 0 {
+		return nil
+	}
 
-		if raw == "" {
-			if err := d.Capture(func(d *jx.Decoder) error {
-				data, err := d.Raw()
-				if err != nil {
-					return err
-				}
-				raw = string(data)
-				return nil
-			}); err != nil {
-				return err
-			}
+	var (
+		raw string
+		got bool
+	)
+	if err := d.Capture(func(d *jx.Decoder) error {
+		data, err := d.Raw()
+		if err != nil {
+			return err
 		}
-
-		e.extract(label, raw)
+		raw = string(data)
+		got = true
+		return nil
+	}); err != nil {
+		return err
+	}
+	if got {
+		for _, label := range n.labels {
+			e.extract(label, raw)
+		}
 	}
 	return nil
 }