--- a/internal/logql/logqlengine/jsonexpr/selector_tree_test.go
+++ b/internal/logql/logqlengine/jsonexpr/selector_tree_test.go
@@ -1,0 +1,11 @@
+package jsonexpr
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestSelectorTree(t *testing.T) {
+	require.True(t, MakeSelectorTree(nil).IsEmpty())
+}