```
goos: linux
goarch: amd64
pkg: github.com/go-faster/oteldb/internal/logql/logqlengine
cpu: AMD Ryzen 9 5950X 16-Core Processor            
                           │   old.txt    │               new.txt               │
                           │    sec/op    │   sec/op     vs base                │
JSONExtractor/JMESPaths-32   2122.0n ± 1%   810.1n ± 1%  -61.82% (p=0.000 n=15)

                           │   old.txt    │                new.txt                │
                           │     B/s      │     B/s       vs base                 │
JSONExtractor/JMESPaths-32   151.9Mi ± 1%   397.9Mi ± 1%  +161.94% (p=0.000 n=15)

                           │  old.txt   │              new.txt               │
                           │    B/op    │    B/op     vs base                │
JSONExtractor/JMESPaths-32   888.0 ± 0%   170.0 ± 0%  -80.86% (p=0.000 n=15)

                           │   old.txt   │              new.txt               │
                           │  allocs/op  │ allocs/op   vs base                │
JSONExtractor/JMESPaths-32   36.000 ± 0%   9.000 ± 0%  -75.00% (p=0.000 n=15)
```